Support only use feature target
