+++ Cargo.toml
@@ -27,10 +27,9 @@ license.workspace = true
 
 [dependencies]
 log = "0.4.11"
-log4rs = { version = "1.1.1", default-features = false, features = ["all_components", "humantime", "serde", "serde-value", "yaml_format", "gzip"] }
-once_cell = "1.15.0"
-serde = { version = "1.0.145", features = ["derive"] }
-is_debug = "1.0.1"
+log4rs = { version = "1.1.1", default-features = false, features = ["all_components", "humantime", "serde", "serde-value", "yaml_format", "gzip"], optional = true }
+once_cell = { version = "1.15.0", default-features = false, optional = true }
+serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 
 [dependencies.simple-log-derive]
 path = "derive"
@@ -47,6 +46,8 @@ members = ["./", "derive"]
 
 
 [features]
+default = ["log_inner"]
+log_inner = ["log4rs", "once_cell"]
 target = ["simple-log-derive"]
 
 [[example]]
